+++ example/android/app/src/main/kotlin/com/smollaitc/update_center_example/MainActivity.kt
@@ -2,5 +2,4 @@ package com.smollaitc.update_center_example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity: FlutterActivity() {}