+++ lib/config/config.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import '../utils/constants.dart';
 
 /// Global Configuration for the plugin
@@ -6,70 +5,27 @@ class GlobalConfig {
   final bool isCheckStart;
   final bool isNoUpdateAvailableToast;
   final bool isSourceUrl;
-  final bool isVerifiedSha256Android;
-  final bool isVerifiedSha256Windows;
   final bool isRequestForNotifications;
   final bool isOpenFile;
-  final bool isEnabledLog;
-  final bool isMarkdownChangelog;
+  final AndroidDialogBuilder? androidDialogBuilder;
+  final IOSDialogBuilder? iosDialogBuilder;
+  final WindowsDialogBuilder? windowsDialogBuilder;
 
+  final NoUpdateAvailableBuilder? androidNoUpdateAvailableBuilder;
+  final NoUpdateAvailableBuilder? iosNoUpdateAvailableBuilder;
+  final NoUpdateAvailableBuilder? windowsNoUpdateAvailableBuilder;
   GlobalConfig({
     this.isCheckStart = false,
     this.isNoUpdateAvailableToast = false,
     this.isSourceUrl = false,
-    this.isVerifiedSha256Android = false,
-    this.isVerifiedSha256Windows = false,
     this.isRequestForNotifications = false,
     this.isOpenFile = false,
-    this.isEnabledLog = false,
-    this.isMarkdownChangelog = false,
-  });
-}
-
-/// UI Configuration for Alert Dialog, Bottom Sheets, and Download Progress
-class UIConfig {
-  final String updateButtonText;
-  final String skipButtonText;
-  final String updateAvailableText;
-  final String changelogText;
-  final String titleDownloadBottomSheets;
-  final String titleVerifiedSha256BottomSheets;
-  final String toastNoUpdateFoundText;
-
-  final TextStyle alertVersionNameStyle;
-  final TextStyle alertChangeLogTextStyle;
-  final TextStyle bottomSheetVersionNameTextStyle;
-  final TextStyle bottomSheetChangeLogTextStyle;
-  final TextStyle updateAvailableTextStyle;
-  final TextStyle changelogTextStyle;
-
-  final Icon customIconTitle;
-  final DialogType dialogType;
-
-  UIConfig({
-    this.updateButtonText = 'Update',
-    this.skipButtonText = 'Skip',
-    this.updateAvailableText = 'Update Available',
-    this.changelogText = 'Changelog',
-    this.titleDownloadBottomSheets = 'Downloading...',
-    this.titleVerifiedSha256BottomSheets = 'Verified sha256...',
-    this.toastNoUpdateFoundText = 'No update found',
-    this.alertVersionNameStyle =
-        const TextStyle(fontSize: 25, fontWeight: FontWeight.bold),
-    this.alertChangeLogTextStyle =
-        const TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
-    this.bottomSheetVersionNameTextStyle =
-        const TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-    this.bottomSheetChangeLogTextStyle =
-        const TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
-    this.updateAvailableTextStyle = const TextStyle(
-      fontSize: 24,
-      fontWeight: FontWeight.bold,
-    ),
-    this.changelogTextStyle =
-        const TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-    this.customIconTitle = const Icon(Icons.downloading_outlined),
-    this.dialogType = DialogType.bottomSheet,
+    this.androidDialogBuilder,
+    this.iosDialogBuilder,
+    this.windowsDialogBuilder,
+    this.androidNoUpdateAvailableBuilder,
+    this.iosNoUpdateAvailableBuilder,
+    this.windowsNoUpdateAvailableBuilder,
   });
 }
 
@@ -80,9 +36,6 @@ class NotificationConfig {
   final String downloadProgressNotificationTextBody;
   final String downloadFailedNotificationTitleText;
   final String downloadFailedNotificationBodyText;
-  final String verifiedSha256NotificationTitleText;
-  final String verifiedSha256NotificationBodyText;
-
   final bool showProgress;
   final bool channelShowBadge;
 
@@ -93,8 +46,6 @@ class NotificationConfig {
     this.downloadFailedNotificationTitleText = 'Download failed',
     this.downloadFailedNotificationBodyText =
         'An error occurred while downloading update. Check your internet connections and try again',
-    this.verifiedSha256NotificationTitleText = 'Verified sha256',
-    this.verifiedSha256NotificationBodyText = '',
     this.showProgress = true,
     this.channelShowBadge = true,
   });
@@ -103,12 +54,10 @@ class NotificationConfig {
 /// Config for the plugin
 class UpdateCenterConfig {
   final GlobalConfig globalConfig;
-  final UIConfig uiConfig;
   final NotificationConfig notificationConfig;
 
   UpdateCenterConfig({
     required this.globalConfig,
-    required this.uiConfig,
     required this.notificationConfig,
   });
 }