+++ lib/models/windows.dart
@@ -1,6 +1,6 @@
 /// Json model for windows
 class WindowsModel {
-  String versionName, downloadUrl, changeLog, sourceUrl, sha256checksum;
+  String versionName, downloadUrl, changeLog, sourceUrl;
   int versionCode, minSupport;
 
   WindowsModel(
@@ -9,7 +9,6 @@ class WindowsModel {
     this.downloadUrl,
     this.changeLog,
     this.sourceUrl,
-    this.sha256checksum,
 
     /// int
     this.versionCode,