+++ lib/provider/check_provider.dart
@@ -1,25 +1,15 @@
 import 'package:flutter/material.dart';
-import 'package:fluttertoast/fluttertoast.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import '../config/config.dart';
 import '../utils/download_utils.dart';
 import '../models/android.dart';
 import '../models/ios.dart';
 import '../models/windows.dart';
-import 'dialog_provider.dart';
 import 'memory_provider.dart';
 
 /// The CheckProvider class is responsible for handling update checks for different platforms.
 class CheckProvider {
   /// Checks for an available update for Android platform and shows update dialog if necessary.
-  ///
-  /// [androidData] - Data containing update details for Android.
-  /// [packageInfo] - Current app package information.
-  /// [allowSkip] - Flag to allow skipping the update.
-  /// [context] - Current BuildContext.
-  /// [downloadState] - State manager for download progress.
-  /// [config] - Configuration settings for the update.
-  /// [downloadUrl] - URL to download the update.
   Future<bool> checkAndroidUpdate(
     Map<String, dynamic> androidData,
     PackageInfo packageInfo,
@@ -34,7 +24,6 @@ class CheckProvider {
       androidData['downloadUrl'],
       androidData['changeLog'],
       androidData['sourceUrl'],
-      androidData['sha256checksum'],
       androidData['versionCode'],
       androidData['minSupport'],
     );
@@ -59,28 +48,30 @@ class CheckProvider {
       // Save the new version info
       await MemoryProvider.saveVersionInfoAndroid(model.versionName);
 
-      if (context.mounted) {
-        DialogProvider().showUpdateDialog(
-            model.versionName,
-            model.changeLog,
-            context,
-            allowSkip,
-            downloadState,
-            model.downloadUrl,
-            model.sha256checksum,
-            model.sourceUrl,
-            config);
+      if (context.mounted && config.globalConfig.androidDialogBuilder != null) {
+        await config.globalConfig.androidDialogBuilder!(
+          context,
+          model, // Assume this is your AndroidModel instance with update info
+          config,
+          downloadState, // Your DownloadState instance
+          allowSkip,
+        );
       }
+
       downloadUrl = model.downloadUrl; // Set the download URL
 
       return true; // Return true to indicate that an update is available
     }
 
+    // Returns your custom widget "No updates found"
     if (buildNumber >= model.versionCode) {
-      if (config.globalConfig.isNoUpdateAvailableToast) {
-        Fluttertoast.showToast(msg: config.uiConfig.toastNoUpdateFoundText);
+      if (context.mounted &&
+          config.globalConfig.androidNoUpdateAvailableBuilder != null) {
+        config.globalConfig.androidNoUpdateAvailableBuilder?.call(context);
       }
-      MemoryProvider.deleteFileDirectory();
+
+      MemoryProvider
+          .deleteFileDirectory(); // Deletes the old file if there is no update
       return false;
     }
 
@@ -88,9 +79,6 @@ class CheckProvider {
   }
 
   /// Checks for an available update for iOS platform and shows update dialog if necessary.
-  ///
-  /// [iosData] - Data containing update details for iOS.
-  /// Similar parameters as checkAndroidUpdate.
   Future<bool> checkIOSUpdate(
     Map<String, dynamic> iosData,
     PackageInfo packageInfo,
@@ -99,7 +87,6 @@ class CheckProvider {
     DownloadState downloadState,
     UpdateCenterConfig config,
     String downloadUrl, // This value is not used in the iOS model
-    String sha256checksum, // This value is not used in the iOS model
   ) async {
     IOSModel model = IOSModel(
       iosData['versionName'],
@@ -117,23 +104,22 @@ class CheckProvider {
 
     // Check if the current build number is less than the update version code
     if (buildNumber < model.versionCode) {
-      DialogProvider().showUpdateDialog(
-        model.versionName,
-        model.changeLog,
-        context,
-        allowSkip,
-        downloadState,
-        downloadUrl,
-        model.sourceUrl,
-        sha256checksum,
-        config,
-      );
+      if (context.mounted && config.globalConfig.iosDialogBuilder != null) {
+        await config.globalConfig.iosDialogBuilder!(
+            context,
+            model, // Assume this is your AndroidModel instance with update info
+            config,
+            downloadState, // Your DownloadState instance
+            allowSkip);
+      }
+
       return true; // Return true to indicate that an update is available
     }
 
     if (buildNumber >= model.versionCode) {
-      if (config.globalConfig.isCheckStart == false) {
-        Fluttertoast.showToast(msg: "No updates found");
+      if (context.mounted &&
+          config.globalConfig.iosNoUpdateAvailableBuilder != null) {
+        config.globalConfig.iosNoUpdateAvailableBuilder?.call(context);
       }
       return false;
     }
@@ -141,8 +127,6 @@ class CheckProvider {
   }
 
   /// Checks for an available update for Windows platform and shows update dialog if necessary.
-  ///
-  /// [windowsData] - Data containing update details for Windows.
   Future<bool> checkWindowsUpdate(
     Map<String, dynamic> windowsData,
     PackageInfo packageInfo,
@@ -157,7 +141,6 @@ class CheckProvider {
       windowsData['downloadUrl'],
       windowsData['changeLog'],
       windowsData['sourceUrl'],
-      windowsData['sha256checksum'],
       windowsData['versionCode'],
       windowsData['minSupport'],
     );
@@ -182,28 +165,26 @@ class CheckProvider {
       // Save the new version info
       await MemoryProvider.saveVersionInfoWindows(model.versionName);
 
-      if (context.mounted) {
-        DialogProvider().showUpdateDialog(
-            model.versionName,
-            model.changeLog,
+      if (context.mounted && config.globalConfig.windowsDialogBuilder != null) {
+        await config.globalConfig.windowsDialogBuilder!(
             context,
-            allowSkip,
-            downloadState,
-            model.downloadUrl,
-            model.sha256checksum,
-            model.sourceUrl,
-            config);
+            model, // Assume this is your AndroidModel instance with update info
+            config,
+            downloadState, // Your DownloadState instance
+            allowSkip);
       }
+
       downloadUrl = model.downloadUrl; // Set the download URL
       return true; // Return true to indicate that an update is available
     }
 
     if (buildNumber >= model.versionCode) {
       if (config.globalConfig.isNoUpdateAvailableToast) {
-        ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-          content: Text(config.uiConfig.toastNoUpdateFoundText),
-          duration: const Duration(seconds: 1),
-        ));
+        if (context.mounted &&
+            config.globalConfig.windowsNoUpdateAvailableBuilder != null) {
+          config.globalConfig.windowsNoUpdateAvailableBuilder?.call(context);
+        }
+        // toast for windows
       }
       MemoryProvider.deleteFileDirectoryWindows();
       return false;