+++ lib/provider/download_provider.dart
@@ -1,11 +1,10 @@
+import 'dart:developer';
 import 'dart:io';
-import 'package:flutter/foundation.dart';
 import 'package:open_filex/open_filex.dart';
 import 'package:http/http.dart' as h;
 import 'package:update_center/provider/memory_provider.dart';
 import 'package:update_center/provider/notification_provider.dart';
 import 'package:update_center/config/config.dart';
-import 'package:update_center/provider/talker_provider.dart';
 import '../utils/download_utils.dart';
 
 /// Provides functionality to handle the download process for updates.
@@ -23,7 +22,6 @@ class DownloadProvider {
     Function(double) onProgress,
     UpdateCenterConfig config,
     DownloadState downloadState,
-    String sha256checksum,
   ) async {
     downloadState.isDownloading.value = true;
 
@@ -76,36 +74,7 @@ class DownloadProvider {
         await fileStream.close();
         notificationProvider.cancelNotification(1000);
 
-        /// Checks the value from the config. If true, then hash check is used.
-        if (config.globalConfig.isVerifiedSha256Android) {
-          downloadState.isVerifiedSha256.value = true;
-
-          notificationProvider.showGenericNotification(
-              id: 4000,
-              title:
-                  config.notificationConfig.verifiedSha256NotificationTitleText,
-              body:
-                  config.notificationConfig.verifiedSha256NotificationBodyText);
-
-          String sha256Result = await compute(computeSha256, file);
-          if (sha256Result == sha256checksum) {
-            // Checksums match
-            await OpenFilex.open(fileName);
-            TalkerProvider(config).talker.info(
-                "Checksum verification successful ${sha256Result.toString()}: used sha Android");
-          } else {
-            // Checksums do not match
-            TalkerProvider(config).talker.info(
-                "Checksum verification failed ${sha256Result.toString()}");
-          }
-          notificationProvider.cancelNotification(4000);
-          downloadState.isVerifiedSha256.value = false;
-        } else {
-          // If a hash check is not used, open the file.
-          await OpenFilex.open(fileName);
-
-          TalkerProvider(config).talker.info("$fileName: unused sha");
-        }
+        await OpenFilex.open(fileName);
 
         downloadState.isDownloading.value = false;
       },
@@ -123,7 +92,7 @@ class DownloadProvider {
                 config.notificationConfig.downloadFailedNotificationTitleText,
             body: config.notificationConfig.downloadFailedNotificationBodyText);
 
-        TalkerProvider(config).talker.error(e);
+        log(e);
       },
       cancelOnError: true,
     );
@@ -136,7 +105,6 @@ class DownloadProvider {
     Function(double) onProgress,
     UpdateCenterConfig config,
     DownloadState downloadState,
-    String sha256checksum,
   ) async {
     downloadState.isDownloading.value = true;
 
@@ -181,34 +149,14 @@ class DownloadProvider {
         await fileStream.flush();
         await fileStream.close();
 
-        // Checks the value from the config. If true, then hash check is used.
-        if (config.globalConfig.isVerifiedSha256Windows) {
-          downloadState.isVerifiedSha256.value = true;
+        await OpenFilex.open(fileName);
 
-          String sha256Result = await compute(computeSha256, file);
-          if (sha256Result == sha256checksum) {
-            // Checksums match
-            await OpenFilex.open(fileName);
-
-            TalkerProvider(config).talker.info(
-                "Checksum verification successful ${sha256Result.toString()}: used sha Windows");
-          } else {
-            // Checksums do not match
-            TalkerProvider(config).talker.info(
-                "Checksum verification failed ${sha256Result.toString()}");
-          }
-          downloadState.isVerifiedSha256.value = false;
-        } else {
-          // If a hash check is not used, open the file.
-          await OpenFilex.open(fileName);
-          TalkerProvider(config).talker.info("$fileName: unused sha");
-        }
         downloadState.isDownloading.value = false;
       },
       onError: (e) async {
         await fileStream.close();
         downloadState.isDownloading.value = false;
-        TalkerProvider(config).talker.error(e);
+        log(e);
       },
       cancelOnError: true,
     );