+++ lib/utils/constants.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
-
-/// Dialogue type
-enum DialogType {
-  alertDialog,
-  bottomSheet,
-}
+import 'package:update_center/config/config.dart';
+import '../models/android.dart';
+import '../models/ios.dart';
+import '../models/windows.dart';
+import 'download_utils.dart';
 
 /// Used to determine the vertical offset
 Widget sizeVer(double height) {
@@ -17,3 +16,26 @@ Widget sizeVer(double height) {
 Widget sizeHor(double width) {
   return SizedBox(width: width);
 }
+
+typedef AndroidDialogBuilder = Future<void> Function(
+    BuildContext context,
+    AndroidModel model,
+    UpdateCenterConfig updateCenterConfig,
+    DownloadState downloadState,
+    bool allowSkip);
+
+typedef IOSDialogBuilder = Future<void> Function(
+    BuildContext context,
+    IOSModel model,
+    UpdateCenterConfig updateCenterConfig,
+    DownloadState downloadState,
+    bool allowSkip);
+
+typedef WindowsDialogBuilder = Future<void> Function(
+    BuildContext context,
+    WindowsModel model,
+    UpdateCenterConfig updateCenterConfig,
+    DownloadState downloadState,
+    bool allowSkip);
+
+typedef NoUpdateAvailableBuilder = void Function(BuildContext context);